+++ src/renderer/src/main.tsx
@@ -9,4 +9,5 @@ import './modern.css'
 import './themes.css'
 import './recap.css'
 import './playing-screen.css'
+import './customization.css'
 createRoot(document.getElementById('root')!).render(<React.StrictMode>{new URLSearchParams(location.search).get('mini') === '1' ? <MiniPlayer/> : <App/>}</React.StrictMode>)